# src/api/specs/swagger.yaml
openapi: 3.0.0
info:
  title: Next.js API
  description: API documentation for Next.js 15 application
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InternalServerError:
      type: object
      properties:
        error:
          type: string
          example: InternalServerError
        message:
          type: string
          example: Server error.
    Success:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Success
    Created:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Created
...
paths:
...
  /api/auth/verify-email:
    get:
      summary: Verify user email
      description: >-
        Verifies the user's email address using a provided verification token.
        Marks the user's email as verified in the database.
      operationId: verifyEmail
      tags:
        - Auth
      parameters:
        - in: query
          name: token
          required: true
          description: The email verification token sent to the user's email address.
          schema:
            type: string
            example: d4c79ab8-b6b7-48b4-b5a4-56e8d41be26g
      responses:
        '200':
          description: Email successfully verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Invalid or expired token, or token not provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Token is required
                  message:
                    type: string
                    example: The provided token is invalid or has expired
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /api/contact:
    post:
      summary: Submit a contact form inquiry
      tags:
        - Contact
      operationId: submitContactForm
      description: Submits a contact form with company details, contact information, and inquiry content. Stores the submission in the database and optionally sends a confirmation email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - companyName
                - contactName
                - email
                - phone
                - inquiry
                - privacyPolicy
              properties:
                companyName:
                  type: string
                  description: The name of the company submitting the contact form.
                  example: Example Corp
                contactName:
                  type: string
                  description: The name of the contact person submitting the form.
                  example: John Doe
                email:
                  type: string
                  format: email
                  description: The email address of the contact person.
                  example: user@example.com
                phone:
                  type: string
                  description: The phone number of the contact person.
                  example: +1234567890
                inquiry:
                  type: string
                  description: The content of the inquiry or message.
                  example: I have a question about your services.
                privacyPolicy:
                  type: boolean
                  description: Indicates whether the user agrees to the privacy policy.
                  example: true
      responses:
        '201':
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
        '400':
          description: Invalid input or missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'User already exists with this email: user@example.com'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'